making config its own package
