+++ model/pyproject.toml
@@ -7,6 +7,7 @@ authors = ["Your Name <you@example.com>"]
 
 [tool.poetry.dependencies]
 python = "^3.12"
+shared = {path = "../config", develop = true}
 scikit-learn = "^1.8.0"
 xgboost-cpu = "^3.1.3"
 pandas = "^3.0.0"
